refactor(settings): extract Appearance section, finish the Display split
Complete the Settings screen restructure (final task of 9): create
AppearanceSection.kt holding Theme, Accent, Theme preset, Map color
(match-app-theme toggle plus the OSM-only color-scheme choices), Glass,
and Font rows, replacing the old ad-hoc Display block in
SettingsScreen.kt. MapThemeMatchRow and the two OSM map-color-scheme
rows move out of MapSection.kt (where Task 7 had temporarily left them)
into this new section, since they read as appearance controls rather
than map-rendering controls.

SettingsScreen.kt now contains only the top-level composable and its
seven section calls, in the final order: Appearance, Screen, Units,
Map, Location, Panels, System.

String changes: add settings_section_appearance and
settings_subheader_map_color; drop the now-unused
settings_section_display and settings_subheader_map_appearance; reword
settings_map_scheme_accent to "Accent (app color)" so the option's
link to the Accent Color control reads directly off its label.

Add two section-smoke tests (renders_appearance_section,
renders_screen_section). Appearance absorbing the map-color rows makes
it taller, which pushes the Screen section's Fullscreen row below the
fold on the short head-unit viewport; scroll it into view in the two
affected pre-existing tests, matching the pattern already used
elsewhere in the same file.
